chg: dev: move core code to module init !refactor
