+++ setup.cfg
@@ -52,6 +52,7 @@ doc =
     sphinxcontrib-apidoc
 test =
     pytest
+    requests
 cov =
     pytest-cov
     coverage[toml]
