+++ tox.ini
@@ -54,7 +54,7 @@ allowlist_externals =
 deps =
     {[base]deps}
     honcho
-    -e .
+    -e .[test]
 
 commands =
     python pyserv/settings.py
